随便更新

+++ app/src/main/java/cc/android/testapp/util/CLog.java
@@ -12,6 +12,7 @@ import java.nio.charset.StandardCharsets;
 import java.text.SimpleDateFormat;
 import java.util.Locale;
 
+import cc.android.testapp.cfg.Config;
 import cc.commons.util.FileUtil;
 import cc.commons.util.IOUtil;
 
@@ -23,7 +24,15 @@ public class CLog {
     private static SimpleDateFormat mTimeF = new SimpleDateFormat("HH:mm:ss.SSS", Locale.CHINESE);
     private static int mCount = 0;
 
+    /**
+     * 日志Tag
+     */
     public static final String TAG = "TestApp";
+
+    /**
+     * 强制输出日志到文件,用于调试用,用完即关
+     */
+    public static boolean FORCE_FILE_OUTPUT = false;
 
     public static void log(String pMsg, Throwable exp) {
         log(pMsg);
@@ -41,19 +50,21 @@ public class CLog {
 
 
     synchronized public static void log(String pLog) {
-        FileOutputStream tFOStream = null;
-        try {
-            if (share_log_file.exists() && (System.currentTimeMillis() - share_log_file.lastModified()) / 1000 > 1200)
-                share_log_file.delete();
-            tFOStream = FileUtil.openOutputStream(share_log_file, true);
-            byte[] tB = String.format(Locale.CHINESE, "%s|%03d: [%s]%s\n",
-                    mTimeF.format(System.currentTimeMillis()), ++mCount,TAG, pLog).getBytes(StandardCharsets.UTF_8);
-            tFOStream.write(tB, 0, tB.length);
-        } catch (IOException e) {
-            e.printStackTrace();
-        } finally {
-            IOUtil.closeStream(tFOStream);
+        if (FORCE_FILE_OUTPUT || Config.enableFilelog()) {
+            FileOutputStream tFOStream = null;
+            try {
+                if (share_log_file.exists() && (System.currentTimeMillis() - share_log_file.lastModified()) / 1000 > 1200)
+                    share_log_file.delete();
+                tFOStream = FileUtil.openOutputStream(share_log_file, true);
+                byte[] tB = String.format(Locale.CHINESE, "%s|%03d: [%s]%s\n",
+                        mTimeF.format(System.currentTimeMillis()), ++mCount, TAG, pLog).getBytes(StandardCharsets.UTF_8);
+                tFOStream.write(tB, 0, tB.length);
+            } catch (IOException e) {
+                Log.e(CLog.TAG, "Error on write log" + e.getLocalizedMessage());
+            } finally {
+                IOUtil.closeStream(tFOStream);
+            }
         }
-        Log.i(TAG,pLog);
+        Log.i(TAG, pLog);
     }
 }
